Allow bookend to only contain a footer
Bookend forces there to be both a header and footer. I needed a bookend that was only a footer. The code previously overwrote the last variable if the firstClass didn't exist. I decoupled the lastClass logic from the firstClass logic so we could have a standalone header bookend and a standalone footer bookend.